# Significantly improve performance of database observers