--- a/Sources/StreamChat/Config/StreamRuntimeCheck.swift
+++ b/Sources/StreamChat/Config/StreamRuntimeCheck.swift
@@ -39,6 +39,8 @@ public enum StreamRuntimeCheck {
     ///  Enables using our legacy web socket connection.
     public static var _useLegacyWebSocketConnection = false
     
+    /// For *internal use* only
+    ///
     /// Enables reusing unchanged converted items in database observers.
-    public static var isDatabaseObserverItemReusingEnabled = true
+    public static var _isDatabaseObserverItemReusingEnabled = true
 }