--- a/Sources/StreamChat/Controllers/PollController/PollVoteListController.swift
+++ b/Sources/StreamChat/Controllers/PollController/PollVoteListController.swift
@@ -194,7 +194,13 @@ extension PollVoteListController {
             _ itemCreator: @escaping (PollVoteDTO) throws -> PollVote
         )
             -> ListDatabaseObserverWrapper<PollVote, PollVoteDTO> = {
-                ListDatabaseObserverWrapper(isBackground: $0, database: $1, fetchRequest: $2, itemCreator: $3)
+                ListDatabaseObserverWrapper(
+                    isBackground: $0,
+                    database: $1,
+                    fetchRequest: $2,
+                    itemCreator: $3,
+                    itemReuseKeyPaths: (\PollVote.id, \PollVoteDTO.id)
+                )
             }
     }
 }