--- a/Sources/StreamChat/Controllers/ThreadListController/ThreadListController.swift
+++ b/Sources/StreamChat/Controllers/ThreadListController/ThreadListController.swift
@@ -204,7 +204,11 @@ extension ChatThreadListController {
         )
             -> ListDatabaseObserverWrapper<ChatThread, ThreadDTO> = {
                 ListDatabaseObserverWrapper(
-                    isBackground: $0, database: $1, fetchRequest: $2, itemCreator: $3
+                    isBackground: $0,
+                    database: $1,
+                    fetchRequest: $2,
+                    itemCreator: $3,
+                    itemReuseKeyPaths: nil // no id
                 )
             }
     }