--- a/Sources/StreamChat/Models/MessageReaction.swift
+++ b/Sources/StreamChat/Models/MessageReaction.swift
@@ -7,6 +7,8 @@ import Foundation
 /// A type representing a message reaction. `ChatMessageReaction` is an immutable snapshot
 /// of a message reaction entity at the given time.
 public struct ChatMessageReaction: Hashable {
+    let id: String
+    
     /// The reaction type.
     public let type: MessageReactionType