--- a/Sources/StreamChat/StateLayer/ChatState+Observer.swift
+++ b/Sources/StreamChat/StateLayer/ChatState+Observer.swift
@@ -43,12 +43,14 @@ extension ChatState {
                     shouldShowShadowedMessages: clientConfig.shouldShowShadowedMessages
                 ),
                 itemCreator: { try $0.asModel() },
+                itemReuseKeyPaths: (\ChatMessage.id, \MessageDTO.id),
                 sorting: []
             )
             watchersObserver = StateLayerDatabaseObserver(
                 databaseContainer: database,
                 fetchRequest: UserDTO.watcherFetchRequest(cid: cid),
                 itemCreator: { try $0.asModel() },
+                itemReuseKeyPaths: (\ChatUser.id, \UserDTO.id),
                 sorting: []
             )
             self.eventNotificationCenter = eventNotificationCenter