… CHANGELOG.md
… Sources/StreamChat/StateLayer/ReactionListState+Observer.swift
… Sources/StreamChat/StateLayer/UserListState+Observer.swift
… Sources/StreamChat/Utils/Database/DatabaseItemConverter.swift
… StreamChat.xcodeproj/project.pbxproj
… TestTools/StreamChatTestTools/Mocks/Models + Extensions/ChatMessageReaction_Mock.swift
… Tests/StreamChatTests/Controllers/ChannelWatcherListController/ChatChannelWatcherListController_Tests.swift
… Tests/StreamChatTests/Controllers/MemberListController/MemberListController_Tests.swift
… Tests/StreamChatTests/Controllers/MessageController/MessageController+Combine_Tests.swift
… Tests/StreamChatTests/Controllers/MessageController/MessageController+SwiftUI_Tests.swift
… Tests/StreamChatTests/Controllers/MessageController/MessageController_Tests.swift
… Tests/StreamChatTests/Controllers/ThreadListController/ChatThreadListController_Tests.swift
… Tests/StreamChatTests/Database/DTOs/MessageDTO_Tests.swift
… Tests/StreamChatTests/ListDatabaseObserverWrapper_Tests.swift
… Tests/StreamChatTests/Query/Sorting/ListDatabaseObserver+Sorting_Tests.swift
… Tests/StreamChatTests/StateLayer/StateLayerDatabaseObserver_Tests.swift
… Tests/StreamChatUITests/SnapshotTests/ChatMessageList/ChatMessage/ChatMessageLayoutOptionsResolver_Tests.swift