### 🔗 Issue Links

Related [ios-issues-tracking/issues/870](https://github.com/GetStream/ios-issues-tracking/issues/870)

### 🎯 Goal

* Improved performance of database observers by reusing unchanged items

### 📝 Summary

* Reuse existing model instances what did not change when NSFetchedResultsController reports list changes
* Reuse newly created items stored in `ListChange` when did change content callback is triggered
* Affects `BackgroundDatabaseObserver` and `StateLayerDatabaseObserver` (list type)

### 🛠 Implementation

`DTO` and immutable model types are matched using a single `id` value. Since `ListChange.item` contains a newly converted item, instead of recreating it again, we reuse the value.

### 🎨 Showcase

When comparing how much CPU time does the `BackgroundDatabaseObserver.updateItems` takes when opening the demo app, navigating to a channel, loading several pages of messages, then we can see significant improvement.

| Before | After |
| ------ | ----- |
|  ~1100ms  |  ~200ms  |

<img width="1886" alt="result" src="https://github.com/GetStream/stream-chat-swift/assets/1469907/ad308db9-2162-4773-9c39-cee574983d5c">

### 🧪 Manual Testing Notes

Since we reuse existing items and do not convert the whole list, we need to be sure, that NSFetchedResultsController does not miss any data changes in more complex relationships. Therefore, we need to do a manual testing round covering many use-cases.

### ☑️ Contributor Checklist

- [x] I have signed the [Stream CLA](https://docs.google.com/forms/d/e/1FAIpQLScFKsKkAJI7mhCr7K9rEIOpqIDThrWxuvxnwUq2XkHyG154vQ/viewform) (required)
- [x] This change follows zero ⚠️ policy (required)
- [x] This change should be manually QAed
- [x] Changelog is updated with client-facing changes
- [x] New code is covered by unit tests
- [x] Comparison screenshots added for visual changes
- [ ] Affected documentation updated (docusaurus, tutorial, CMS)

### 🎁 Meme

![gif](https://i.giphy.com/media/v1.Y2lkPTc5MGI3NjExNzZtNWp0NzBwdnVidzN6OTI3enR6aWs0a2RpOGZibWx1NGkxN252MiZlcD12MV9pbnRlcm5hbF9naWZfYnlfaWQmY3Q9Zw/bTMzzPaOABxzG/giphy.gif)